Remove question

+++ MivBitstream/src/SeiRBSP.cpp
@@ -154,7 +154,7 @@ void encodeSeiHeaderValue(std::ostream &stream, size_t value) {
 } // namespace
 
 void SeiMessage::encodeTo(std::ostream &stream) const {
-  encodeSeiHeaderValue(stream, unsigned(payloadType()));  // TODO why does even the method use accessors and not directly the private variables?
+  encodeSeiHeaderValue(stream, static_cast<std::uint64_t>(payloadType()));
   encodeSeiHeaderValue(stream, payloadSize());
   stream.write(payload().data(), payload().size());
 }
